Implement migrate groups and move device migrate functions to the device module
