+++ crml/sylo/src/groups/tests.rs
@@ -18,7 +18,7 @@ mod tests {
 	use crate::groups::{AcceptPayload, Encode, Error, Group, Invite, Member, MemberRoles, Module};
 	use crate::mock::{ExtBuilder, Origin, Test};
 	use crate::vault;
-	use frame_support::{assert_err, assert_ok, dispatch::DispatchError};
+	use frame_support::{assert_err, assert_ok};
 	use sp_core::{ed25519, Pair, H256};
 
 	type Groups = Module<Test>;
