+++ crml/sylo/src/migration.rs
@@ -13,21 +13,11 @@
 *     https://centrality.ai/licenses/lgplv3.txt
 */
 
-use crate::{
-	device::{self, DeviceId},
-	groups, inbox, vault,
-};
-use frame_support::{decl_error, decl_module, decl_storage, dispatch::Vec, ensure, weights::SimpleDispatchInfo};
+use frame_support::{decl_module, decl_storage, ensure, weights::SimpleDispatchInfo};
 use frame_system::{ensure_root, ensure_signed};
 use sp_runtime::{DispatchError::BadOrigin, DispatchResult};
 
-pub trait Trait: device::Trait + groups::Trait + inbox::Trait + vault::Trait {}
-
-decl_error! {
-	pub enum Error for Module<T: Trait> {
-		MaxDeviceLimitReached,
-	}
-}
+pub trait Trait: frame_system::Trait {}
 
 decl_storage! {
 	trait Store for Module<T: Trait> as SyloMigration {
@@ -37,7 +27,6 @@ decl_storage! {
 
 decl_module! {
 	pub struct Module<T: Trait> for enum Call where origin: T::Origin, system = frame_system {
-		type Error = Error<T>;
 
 		#[weight = SimpleDispatchInfo::FixedOperational(0)]
 		pub fn set_migrator_account(origin, account_id: T::AccountId) -> DispatchResult {
@@ -50,25 +39,6 @@ decl_module! {
 		pub fn self_destruct(origin) -> DispatchResult {
 			Self::ensure_sylo_migrator(origin)?;
 			MigrationAccount::<T>::kill();
-			Ok(())
-		}
-
-		#[weight = SimpleDispatchInfo::FixedOperational(0)]
-		fn migrate_devices(origin, user_id: T::AccountId, device_ids: Vec<DeviceId>) -> DispatchResult {
-			Self::ensure_sylo_migrator(origin)?;
-			ensure!(device_ids.len() <= device::MAX_DEVICES, Error::<T>::MaxDeviceLimitReached);
-
-			let mut devices = <device::Devices<T>>::get(user_id.clone());
-			ensure!(devices.len() + device_ids.len() <= device::MAX_DEVICES, Error::<T>::MaxDeviceLimitReached);
-
-			for device_id in device_ids {
-				if !devices.contains(&device_id) {
-					devices.push(device_id);
-				}
-			}
-
-			<device::Devices<T>>::insert(user_id, devices);
-
 			Ok(())
 		}
 	}
@@ -90,9 +60,9 @@ mod tests {
 	use sp_core::H256;
 	use sp_runtime::DispatchError::BadOrigin;
 
-	impl Trait for Test {}
 	type Migration = Module<Test>;
-	type Device = device::Module<Test>;
+
+	impl Trait for Test {}
 
 	#[test]
 	fn set_migration_account_works() {
@@ -162,65 +132,6 @@ mod tests {
 			);
 
 			assert_ok!(Migration::ensure_sylo_migrator(Origin::signed(migration_account)));
-		});
-	}
-
-	#[test]
-	fn migrate_devices_works() {
-		ExtBuilder::default().build().execute_with(|| {
-			let user_id = H256::from_low_u64_be(1);
-			let devices = vec![1, 2, 3, 4];
-			let migration_account = H256::from_low_u64_be(2);
-
-			assert_ok!(Migration::set_migrator_account(Origin::ROOT, migration_account));
-
-			assert_ok!(Migration::migrate_devices(
-				Origin::signed(migration_account),
-				user_id.clone(),
-				devices.clone()
-			));
-
-			assert_eq!(Device::devices(user_id), [1, 2, 3, 4]);
-		});
-	}
-
-	#[test]
-	fn migrate_devices_does_not_double_up() {
-		ExtBuilder::default().build().execute_with(|| {
-			let user_id = H256::from_low_u64_be(1);
-			let devices = vec![1, 2, 3, 4];
-			let migration_account = H256::from_low_u64_be(2);
-
-			assert_ok!(Migration::set_migrator_account(Origin::ROOT, migration_account));
-
-			assert_ok!(Device::append_device(&user_id, 3));
-
-			assert_ok!(Migration::migrate_devices(
-				Origin::signed(migration_account),
-				user_id.clone(),
-				devices.clone()
-			));
-
-			assert_eq!(Device::devices(user_id), [3, 1, 2, 4]);
-		});
-	}
-
-	#[test]
-	fn migrate_devices_fails_with_bad_account() {
-		ExtBuilder::default().build().execute_with(|| {
-			let user_id = H256::from_low_u64_be(1);
-			let devices = vec![1, 2, 3, 4];
-			let migration_account = H256::from_low_u64_be(2);
-			let invalid_account = H256::from_low_u64_be(3);
-
-			assert_ok!(Migration::set_migrator_account(Origin::ROOT, migration_account));
-
-			assert_eq!(
-				Migration::migrate_devices(Origin::signed(invalid_account), user_id.clone(), devices.clone()),
-				Err(BadOrigin)
-			);
-
-			assert_eq!(Device::devices(user_id), []);
 		});
 	}
 }
